+++ cypress/integration/abc.spec.js
@@ -5,10 +5,20 @@
 *
 * Let's test the library separately from the demos (allows us to detach them).
 */
+import {createEventDispatcher} from "svelte";
+
 describe('svg.rx.js', () => {
     beforeEach(() => {
         cy.visit('index.html')
     });
+
+    // tbd. Iterate over the SVG's and send 'cyTest' event to them.
+    //
+
+    let dispatch = createEventDispatcher();
+    function myCustom() {
+        cy.get('svg#case-1').dispatch("cyTest");
+    }
 
     it('has an SVG cradle', () => {     // placeholder
         cy.get('svg')