+++ test/cases/Collision-1.svelte
@@ -8,6 +8,8 @@
 <script>
     export const blah = 0;		// Testing out properties: <Collision0 blah=... />
 
+    import { onMount } from 'svelte';
+
     import "../../src/main.js";
 
     // tbd. Mocha, Chai...
@@ -17,17 +19,35 @@
     //
     import "chai/register-should";      // https://www.chaijs.com/guide/styles/#should
 
-    const rect0 = document.querySelector("#case-1 > rect");
-    const circle0 = document.querySelector("#case-1 > circle");
-
     let circleClicked = false;
+
+    /*** remove
+    if (typeof describe !== "undefined") {     // running in Cypress (not interactive 'npm run test:dev')
+    ***/
+
+    // Tests are run if opened in Cypress (not the interactive 'npm run test:dev')
+    //
+    function cyTest(cy, describe, it) {
+        describe('#indexOf()', function() {
+            it('should return -1 when the value is not present', function() {
+                assert.equal([1, 2, 3].indexOf(4), -1);
+            });
+        });
+    }
+
+    let svgEl;
+
+    onMount(() => {
+        //const rect0 = document.querySelector("#case-1 > rect");
+        //const circle0 = document.querySelector("#case-1 > circle");
+    });
 </script>
 
 <style>
     circle {
         fill: blue;
         stroke: gray;
-        transition: 300ms;  // any change of any 'circle' parameters (we only change the fill)
+        transition: 300ms;  /* any change of any 'circle' parameters (we only change the fill) */
     }
     circle.selected {
         fill: red;
@@ -35,7 +55,8 @@
 </style>
 
 <h1>Rect with circle (sounds like an art title)</h1>
-<svg viewBox="0 0 220 100" id="case-0">
+<svg viewBox="0 0 220 100" bind:this={svgEl} on:cyTest={cyTest}>
     <rect x="100" y="150" width="100" height="50"></rect>
     <circle class:selected={circleClicked} cx="78" cy="52" r="30" on:click={() => circleClicked = !circleClicked }/>
 </svg>
+<button>Test</button>